requirements.txt
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
exceptiongroup==1.1.3
google-api-core==2.11.1
google-auth==2.22.0
google-cloud-core==2.3.3
google-cloud-storage==2.10.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.60.0
idna==3.4
iniconfig==2.0.0
packaging==23.1
pluggy==1.3.0
protobuf==4.24.2
py-data-mock==0.0.15
pyasn1==0.5.0
pyasn1-modules==0.3.0
pytest==7.4.0
requests==2.31.0
rsa==4.9
six==1.16.0
tomli==2.0.1
urllib3==1.26.16
